+++ projects/angular-grid-layout/src/lib/grid.module.ts
@@ -5,19 +5,22 @@ import { KtdGridItemComponent } from './grid-item/grid-item.component';
 import { KtdGridDragHandle } from './directives/drag-handle';
 import { KtdGridResizeHandle } from './directives/resize-handle';
 import { KtdGridService } from './grid.service';
+import { KtdGridItemPlaceholder } from '../public-api';
 
 @NgModule({
     declarations: [
         KtdGridComponent,
         KtdGridItemComponent,
         KtdGridDragHandle,
-        KtdGridResizeHandle
+        KtdGridResizeHandle,
+        KtdGridItemPlaceholder
     ],
     exports: [
         KtdGridComponent,
         KtdGridItemComponent,
         KtdGridDragHandle,
-        KtdGridResizeHandle
+        KtdGridResizeHandle,
+        KtdGridItemPlaceholder
     ],
     providers: [
         KtdGridService