+++ projects/angular-grid-layout/src/public-api.ts
@@ -2,8 +2,10 @@
  * Public API Surface of grid
  */
 export { ktdGridCompact, ktdTrackById } from './lib/utils/grid.utils';
+export { KtdClientRect } from './lib/utils/client-rect';
 export * from './lib/directives/drag-handle';
 export * from './lib/directives/resize-handle';
+export * from './lib/directives/placeholder';
 export * from './lib/grid-item/grid-item.component';
 export * from './lib/grid.definitions';
 export * from './lib/grid.component';