+++ projects/demo-app/src/app/real-life-example/real-life-example.component.ts
@@ -1,9 +1,9 @@
-import { Component, Inject, OnDestroy, OnInit, QueryList, ViewChild, ViewChildren } from '@angular/core';
-import { KtdGridComponent, KtdGridLayout, ktdTrackById } from '@katoid/angular-grid-layout';
+import { ChangeDetectorRef, Component, Inject, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { KtdGridComponent, KtdGridItemResizeEvent, KtdGridLayout, ktdTrackById } from '@katoid/angular-grid-layout';
 import { fromEvent, merge, Subscription } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
 import { countriesPopulation, countriesPopulationByYear } from './data/countries-population.data';
-import { AreaChartStackedComponent } from '@swimlane/ngx-charts';
+import { Color, LegendPosition, ScaleType } from '@swimlane/ngx-charts';
 import { DOCUMENT } from '@angular/common';
 
 @Component({
@@ -13,22 +13,21 @@ import { DOCUMENT } from '@angular/common';
 })
 export class KtdRealLifeExampleComponent implements OnInit, OnDestroy {
     @ViewChild(KtdGridComponent, {static: true}) grid: KtdGridComponent;
-    @ViewChildren(AreaChartStackedComponent) areaCharts: QueryList<AreaChartStackedComponent>;
 
     trackById = ktdTrackById;
     cols = 12;
     rowHeight = 50;
     compactType: 'vertical' | 'horizontal' | null = 'vertical';
-    layout = [
-        {id: '0', x: 0, y: 5, w: 4, h: 10},
-        {id: '1', x: 4, y: 5, w: 4, h: 10},
-        {id: '2', x: 2, y: 0, w: 6, h: 5},
-        {id: '5', x: 8, y: 0, w: 4, h: 5},
-        {id: '3', x: 0, y: 0, w: 2, h: 5},
-        {id: '4', x: 8, y: 5, w: 4, h: 10}
+    layout: KtdGridLayout = [
+        {id: '0', x: 0, y: 5, w: 4, h: 10, minW: 2, minH: 5},
+        {id: '1', x: 4, y: 5, w: 4, h: 10, minW: 2, minH: 5},
+        {id: '2', x: 2, y: 0, w: 6, h: 5, minW: 4, minH: 4, maxW: 8, maxH: 14},
+        {id: '5', x: 8, y: 0, w: 4, h: 5, minW: 2, minH: 3},
+        {id: '3', x: 0, y: 0, w: 2, h: 5, minH: 3},
+        {id: '4', x: 8, y: 5, w: 4, h: 10, minW: 3, minH: 5, maxH: 12}
     ];
 
-    layoutSizes: {[id: string]: [number, number]} = {};
+    layoutSizes: { [id: string]: [number, number] } = {};
 
 
     countriesPopulation: any[] = countriesPopulation;
@@ -46,22 +45,34 @@ export class KtdRealLifeExampleComponent implements OnInit, OnDestroy {
     yAxisLabel: string = 'Population';
     timeline: boolean = true;
 
-    colorScheme = {
-        domain: ['#5AA454', '#E44D25', '#CFC0BB', '#7aa3e5', '#a8385d', '#aae3f5']
+    colorScheme: Color = {
+        domain: ['#5AA454', '#E44D25', '#CFC0BB', '#7aa3e5', '#a8385d', '#aae3f5'],
+        group: ScaleType.Ordinal,
+        selectable: true,
+        name: 'Customer Usage'
     };
 
-    colorScheme2 = {
-        domain: ['#4e79a7', '#f28e2c', '#e15759', '#76b7b2', '#59a14f', '#edc949', '#af7aa1']
+    colorScheme2: Color = {
+        domain: ['#4e79a7', '#f28e2c', '#e15759', '#76b7b2', '#59a14f', '#edc949', '#af7aa1'],
+        group: ScaleType.Ordinal,
+        selectable: true,
+        name: 'vertical stacked'
     };
 
-    colorSchemeGradientLinear = {
-        domain: ['#4e79a7', '#f28e2c', '#e15759']
+    colorSchemeGradientLinear: Color = {
+        domain: ['#4e79a7', '#f28e2c', '#e15759'],
+        group: ScaleType.Ordinal,
+        selectable: true,
+        name: 'color gradient'
     };
+
+    legendBelow = LegendPosition.Below;
+    schemeType = ScaleType.Linear;
 
 
     private resizeSubscription: Subscription;
 
-    constructor(@Inject(DOCUMENT) public document: Document) { }
+    constructor(@Inject(DOCUMENT) public document: Document, private cd: ChangeDetectorRef) { }
 
     ngOnInit() {
         this.resizeSubscription = merge(
@@ -106,6 +117,12 @@ export class KtdRealLifeExampleComponent implements OnInit, OnDestroy {
 
     onDeactivate(id: string, data): void {
         console.log('Deactivate', JSON.parse(JSON.stringify(data)));
+    }
+
+    onGridItemResize(gridItemResizeEvent: KtdGridItemResizeEvent) {
+        console.log('onGridItemResize', gridItemResizeEvent);
+        this.layoutSizes[gridItemResizeEvent.gridItemRef.id] = [gridItemResizeEvent.width, gridItemResizeEvent.height];
+        this.cd.detectChanges();
     }
 
     /**